add 'mapper', more functions for 'service' and 'controller' packages, fix some mistakes

+++ src/main/java/ru/skypro/homework/service/AdService.java
@@ -2,30 +2,57 @@ package ru.skypro.homework.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.skypro.homework.dto.CreateOrUpdateAdDTO;
+import ru.skypro.homework.dto.AdDTO;
+import ru.skypro.homework.mapper.EntityMapper;
 import ru.skypro.homework.model.AdEntity;
 import ru.skypro.homework.repository.AdRepository;
+
+import java.util.List;
+
 
 @Service
 public class AdService {
     @Autowired
     private AdRepository adRepository;
+    @Autowired
+    private EntityMapper mapper;
 
-    public AdEntity addNewAd(CreateOrUpdateAdDTO ad) {
-        AdEntity newAd = new AdEntity();
-        newAd.setPrice(ad.getPrice());
-        newAd.setTitle(ad.getTitle());
-        newAd.setDescription(ad.getDescription());
-        newAd.setPhoto(newAd.getPhoto());
-        return adRepository.save(newAd);
+
+    public AdDTO getAd(Long id) {
+        AdEntity adEntity = adRepository.findById(id);
+        return mapper.adEntityToAdDTO(adEntity);
     }
 
-    public AdEntity updateInfoAd(CreateOrUpdateAdDTO ad){
-        adRepository.findByTitle(ad.getTitle());
-        return adRepository.updateAd(ad);
+    public long createAd(AdDTO adDTO) {
+        AdEntity adEntity = mapper.adDTOToAdEntity(adDTO);
+        AdEntity savedAd = adRepository.save(adEntity);
+        mapper.adEntityToAdDTO(savedAd);
+        return savedAd.getId();
     }
 
-    public AdEntity deleteAd(long id){
-        return adRepository.deleteById(id);
+    public AdDTO updateInfoAd(long id, AdDTO ad) {
+        AdEntity existingAd = adRepository.findById(id);
+
+        existingAd.setTitle(ad.getTitle());
+        existingAd.setDescription(ad.getDescription());
+        existingAd.setPrice(ad.getPrice());
+        existingAd.setPhoto(ad.getPhoto());
+        existingAd.setComments(ad.getComments());
+
+        AdEntity updatedAd = adRepository.save(existingAd);
+        return mapper.adEntityToAdDTO(updatedAd);
+    }
+
+    public void deleteAd(Long id) {
+        if (!adRepository.existsById(id)) {
+            throw new RuntimeException("Ad not found with id: " + id);
+        }
+        adRepository.deleteById(id);
+    }
+    public List<AdEntity> findAllAds(){
+        return adRepository.findAll();
+    }
+    public List<AdEntity> findAdsOfUser(long id){
+        return adRepository.findByOwnerId(id);
     }
 }
